Migrate to LiquiBase for Database Schema Evolution Management.
This replaces all the sql files for creating and updating a MySQL
database. It also eliminates the need for manual updating of the
database before upgrading the war file. All is now done from within the
application itself. This makes deploying and upgrading as easy as
dropping in a new war file, replacing the old one!