Finished implementing the JdoPersonDAO.
Needed to change the implementation of TeacherAttribute, to make it
flat. Else it is too difficult, when we can't use super-class as
inheritance strategy.

Also deleted the teachus-test project. It needs to be redone after all
these changes, and against our upcoming services instead.

teachus-frontend doesn't compile after the TeacherAttribute change.